requirements.txt
openai
customtkinter
pillow
requests
mutagen
gspread
google-auth
kokoro
soundfile
numpy
opencv-python
faster-whisper
